fix: Remove PUT method from exposed port configuration
