+++ apps/vs-agent/src/services/FullTailsFileService.ts
@@ -1,10 +1,9 @@
 import type { AnonCredsRevocationRegistryDefinition } from '@credo-ts/anoncreds'
-import type { AgentContext } from '@credo-ts/core'
+import type { AgentContext, Logger } from '@credo-ts/core'
 
 import { BasicTailsFileService } from '@credo-ts/anoncreds'
 import { utils } from '@credo-ts/core'
-import axios from 'axios'
-import FormData from 'form-data'
+import { createHash } from 'crypto'
 import fs from 'fs'
 
 export class FullTailsFileService extends BasicTailsFileService {
@@ -24,17 +23,62 @@ export class FullTailsFileService extends BasicTailsFileService {
     const localTailsFilePath = revocationRegistryDefinition.value.tailsLocation
 
     const tailsFileId = utils.uuid()
-    const data = new FormData()
-    const readStream = fs.createReadStream(localTailsFilePath)
-    data.append('file', readStream)
-    const response = await axios.put(`${this.tailsServerBaseUrl}/${encodeURIComponent(tailsFileId)}`, data, {
-      headers: {
-        ...data.getHeaders(),
-      },
-    })
-    if (response.status !== 200) {
-      throw new Error('Cannot upload tails file')
+    try {
+      await processTailsFile(localTailsFilePath, tailsFileId, agentContext.config.logger)
+      console.log('Tails file processed successfully!')
+    } catch (error) {
+      console.error(`Failed to process tails file: ${error.message}`)
     }
     return { tailsFileUrl: `${this.tailsServerBaseUrl}/${encodeURIComponent(tailsFileId)}` }
   }
+}
+
+export const baseFilePath = './tails'
+const indexFilePath = `./${baseFilePath}/index.json`
+
+if (!fs.existsSync(baseFilePath)) {
+  fs.mkdirSync(baseFilePath, { recursive: true })
+}
+export const tailsIndex = (
+  fs.existsSync(indexFilePath) ? JSON.parse(fs.readFileSync(indexFilePath, { encoding: 'utf-8' })) : {}
+) as Record<string, string>
+
+function fileHash(filePath: string, algorithm = 'sha256') {
+  return new Promise<string>((resolve, reject) => {
+    const shasum = createHash(algorithm)
+    try {
+      const s = fs.createReadStream(filePath)
+      s.on('data', function (data) {
+        shasum.update(data)
+      })
+      // making digest
+      s.on('end', function () {
+        const hash = shasum.digest('hex')
+        return resolve(hash)
+      })
+    } catch (error) {
+      return reject('error in calculation')
+    }
+  })
+}
+
+async function processTailsFile(localFilePath: string, tailsFileId: string, logger: Logger) {
+  logger.info(`Processing tails file: ${tailsFileId}`)
+
+  if (!localFilePath) throw new Error('No file path was provided.')
+  if (!tailsFileId) throw new Error('Missing tailsFileId')
+  if (tailsIndex[tailsFileId]) throw new Error(`There is already an entry for: ${tailsFileId}`)
+
+  const hash = await fileHash(localFilePath)
+  const destinationPath = `${baseFilePath}/${hash}`
+  if (fs.existsSync(destinationPath)) {
+    logger.warn('Tails file already exists')
+  } else {
+    fs.copyFileSync(localFilePath, destinationPath)
+  }
+
+  tailsIndex[tailsFileId] = hash
+  fs.writeFileSync(indexFilePath, JSON.stringify(tailsIndex))
+
+  logger.info(`Successfully processed tails file ${tailsFileId}`)
 }
